e-puck 2: Add modal to connect to Wi-Fi
In the modal, the user can enter an IP address to connect to. When the
user wants to disconnect and return to the simulation, they can select
Simulation in the menu.

This has been partially tested with a real e-puck 2, and at least
connecting works. Bluetooth does not work at all.